requirements.txt
livekit-agents>=0.11.2
livekit-plugins-openai>=0.10.6
livekit-plugins-deepgram>=0.6.9
livekit-plugins-silero>=0.7.3
python-dotenv~=1.0
aiofile